trim-whitespace, and fix some eclipse warnings
git-svn-id: https://svn.apache.org/repos/asf/pivot/trunk@1624277 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/org/apache/pivot/util/Version.java
@@ -105,7 +105,7 @@ public class Version implements Comparable<Version>, Serializable {
     @Override
     public String toString() {
         String string = this.majorRevision + "." + this.minorRevision + "."
-            + this.maintenanceRevision + "_" + String.format("%02d", this.updateRevision);
+            + this.maintenanceRevision + "_" + String.format("%02d", Byte.valueOf(this.updateRevision));
 
         if (this.build != null) {
             string += "-" + this.build;
